cacheroach: Add hero diagram
This change adds a diagram generator to the gendoc command. It uses the
go-diagrams library to generate a graphviz file, which is then rendered by the
system installation of graphviz.

When go 1.16 lands, it would be nice to use the static-file import feature to
push the documentation into the bootstrap environment.