cacheroach: Add OIDC user provisioning.
This change allows cacheroach use an OpenID Connect provider to automatically
provision users and removes password-based authentication from cacheroach.
OIDC-derived Principals will use the "offline" authentication flow, allowing
for ongoing revalidation of the upstream user account.

It is still possible to create principals that do not have a backing OIDC
account to use as role accounts for automation.

This change also allows email domains to be used as a means of granting access
to a group of users. This can be seen as an initial step towards #3.

Closes #2.